load version information from VERSION file or use last git sha
